Move `common` back to `cmd`
