+++ cmd/gofer/main.go
@@ -19,19 +19,19 @@ import (
 	"os"
 
 	suite "github.com/chronicleprotocol/oracle-suite"
-	"github.com/chronicleprotocol/oracle-suite/cmd/common"
+	"github.com/chronicleprotocol/oracle-suite/cmd"
 	gofer "github.com/chronicleprotocol/oracle-suite/pkg/config/gofernext"
 )
 
 func main() {
 	var config gofer.Config
-	cf := common.ConfigFlagsForConfig(config)
+	cf := cmd.ConfigFlagsForConfig(config)
 
-	var lf common.LoggerFlags
-	c := common.NewRootCommand("gofer", suite.Version, &cf, &lf)
+	var lf cmd.LoggerFlags
+	c := cmd.NewRootCommand("gofer", suite.Version, &cf, &lf)
 
 	c.AddCommand(
-		common.NewRunCmd(&config, &cf, &lf),
+		cmd.NewRunCmd(&config, &cf, &lf),
 		NewModelsCmd(&config, &cf, &lf),
 		NewDataCmd(&config, &cf, &lf),
 	)
