+++ cmd/spire/cmd_pull.go
@@ -25,11 +25,11 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/chronicleprotocol/oracle-suite/cmd/common"
+	"github.com/chronicleprotocol/oracle-suite/cmd"
 	"github.com/chronicleprotocol/oracle-suite/pkg/config/spire"
 )
 
-func NewPullCmd(c *spire.Config, f *common.ConfigFlags, l *common.LoggerFlags) *cobra.Command {
+func NewPullCmd(c *spire.Config, f *cmd.ConfigFlags, l *cmd.LoggerFlags) *cobra.Command {
 	cc := &cobra.Command{
 		Use:   "pull",
 		Args:  cobra.ExactArgs(1),
@@ -42,7 +42,7 @@ func NewPullCmd(c *spire.Config, f *common.ConfigFlags, l *common.LoggerFlags) *
 	return cc
 }
 
-func NewPullPriceCmd(c *spire.Config, f *common.ConfigFlags, l *common.LoggerFlags) *cobra.Command {
+func NewPullPriceCmd(c *spire.Config, f *cmd.ConfigFlags, l *cmd.LoggerFlags) *cobra.Command {
 	return &cobra.Command{
 		Use:   "price ASSET_PAIR FEED",
 		Args:  cobra.ExactArgs(2),
@@ -87,7 +87,7 @@ type pullPricesOptions struct {
 	FilterFrom string
 }
 
-func NewPullPricesCmd(c *spire.Config, f *common.ConfigFlags, l *common.LoggerFlags) *cobra.Command {
+func NewPullPricesCmd(c *spire.Config, f *cmd.ConfigFlags, l *cmd.LoggerFlags) *cobra.Command {
 	var pullPricesOpts pullPricesOptions
 	cc := &cobra.Command{
 		Use:   "prices",
