+++ cmd/test/testdata/stage-arb-goerli.txt
