+++ cmd/test/testdata/stage-gno.txt
