+++ cmd/test/testdata/stage-gor.txt
