+++ cmd/test/testdata/stage-mango.txt
