+++ cmd/test/testdata/stage-ogor.txt
