+++ cmd/test/testdata/stage-sep.txt
