+++ cmd/test/testdata/stage.txt
